common_wal

Module options

Re-exports

pub use crate::options::kafka::KafkaWalOptions;

Modules

kafka

kafka_prefix ๐

Enums

WalOptions

Wal options allocated to a region. A wal options is encoded by metasrv with serde_json::to_string, and then decoded by datanode with serde_json::from_str.

Constants

WAL_OPTIONS_KEY

An encoded wal options will be wrapped into a (WAL_OPTIONS_KEY, encoded wal options) key-value pair and inserted into the options of a RegionCreateRequest.